fix the skeleton
By default the triclops quickapp project skeleton generator makes a
hash of the formatting of the projects asdf file, and the Makefile
does not compress the binary, because only SBCL seems to support this
use. Since I am targeting SBCL for binary deployment, I have added the
flag to compress the binary when it is generated by the Makefile.